+++ internal/provider/model_resource_meraki_sm_admin_roles.go
@@ -1,0 +1,200 @@
+// Copyright © 2024 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+	"net/url"
+
+	"github.com/CiscoDevNet/terraform-provider-meraki/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/netascode/go-meraki"
+	"github.com/tidwall/gjson"
+	"github.com/tidwall/sjson"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin types
+
+type ResourceSMAdminRoles struct {
+	Id             types.String                `tfsdk:"id"`
+	OrganizationId types.String                `tfsdk:"organization_id"`
+	Items          []ResourceSMAdminRolesItems `tfsdk:"items"`
+}
+
+type ResourceSMAdminRolesItems struct {
+	Id    types.String `tfsdk:"id"`
+	Name  types.String `tfsdk:"name"`
+	Scope types.String `tfsdk:"scope"`
+	Tags  types.List   `tfsdk:"tags"`
+}
+
+// End of section. //template:end types
+
+// Section below is generated&owned by "gen/generator.go". //template:begin getPath
+
+func (data ResourceSMAdminRoles) getPath() string {
+	return fmt.Sprintf("/organizations/%v/sm/admins/roles", url.QueryEscape(data.OrganizationId.ValueString()))
+}
+
+// End of section. //template:end getPath
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toBody
+
+func (data ResourceSMAdminRolesItems) toBody(ctx context.Context, state ResourceSMAdminRolesItems) string {
+	body := ""
+	if !data.Name.IsNull() {
+		body, _ = sjson.Set(body, "name", data.Name.ValueString())
+	}
+	if !data.Scope.IsNull() {
+		body, _ = sjson.Set(body, "scope", data.Scope.ValueString())
+	}
+	if !data.Tags.IsNull() {
+		var values []string
+		data.Tags.ElementsAs(ctx, &values, false)
+		body, _ = sjson.Set(body, "tags", values)
+	}
+	return body
+}
+
+// End of section. //template:end toBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBody
+
+func (data *ResourceSMAdminRoles) fromBody(ctx context.Context, res meraki.Res) {
+	data.Items = make([]ResourceSMAdminRolesItems, 0)
+	res.ForEach(func(k, res gjson.Result) bool {
+		parent := &data
+		data := ResourceSMAdminRolesItems{}
+		if value := res.Get("name"); value.Exists() && value.Value() != nil {
+			data.Name = types.StringValue(value.String())
+		} else {
+			data.Name = types.StringNull()
+		}
+		if value := res.Get("scope"); value.Exists() && value.Value() != nil {
+			data.Scope = types.StringValue(value.String())
+		} else {
+			data.Scope = types.StringNull()
+		}
+		if value := res.Get("tags"); value.Exists() && value.Value() != nil {
+			data.Tags = helpers.GetStringList(value.Array())
+		} else {
+			data.Tags = types.ListNull(types.StringType)
+		}
+		(*parent).Items = append((*parent).Items, data)
+		return true
+	})
+}
+
+// End of section. //template:end fromBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyPartial
+
+// fromBodyPartial reads values from a gjson.Result into a tfstate model. It ignores null attributes in order to
+// uncouple the provider from the exact values that the backend API might summon to replace nulls. (Such behavior might
+// easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
+// "managed" elements, instead of all elements.
+func (data *ResourceSMAdminRoles) fromBodyPartial(ctx context.Context, res meraki.Res) {
+	for i := range data.Items {
+		parent := &data
+		data := (*parent).Items[i]
+		parentRes := &res
+		var res gjson.Result
+
+		parentRes.ForEach(
+			func(_, v gjson.Result) bool {
+				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+					res = v
+					return false
+				}
+				return true
+			},
+		)
+		if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
+			data.Name = types.StringValue(value.String())
+		} else {
+			data.Name = types.StringNull()
+		}
+		if value := res.Get("scope"); value.Exists() && !data.Scope.IsNull() {
+			data.Scope = types.StringValue(value.String())
+		} else {
+			data.Scope = types.StringNull()
+		}
+		if value := res.Get("tags"); value.Exists() && !data.Tags.IsNull() {
+			data.Tags = helpers.GetStringList(value.Array())
+		} else {
+			data.Tags = types.ListNull(types.StringType)
+		}
+	}
+}
+
+// End of section. //template:end fromBodyPartial
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyUnknowns
+
+// fromBodyUnknowns updates the Unknown Computed tfstate values from a JSON.
+// Known values are not changed (usual for Computed attributes with UseStateForUnknown or with Default).
+func (data *ResourceSMAdminRoles) fromBodyUnknowns(ctx context.Context, res meraki.Res) {
+}
+
+// End of section. //template:end fromBodyUnknowns
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toDestroyBody
+
+func (data ResourceSMAdminRoles) toDestroyBody(ctx context.Context) string {
+	body := ""
+	return body
+}
+
+// End of section. //template:end toDestroyBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin hasChanges
+
+func (data *ResourceSMAdminRoles) hasChanges(ctx context.Context, state *ResourceSMAdminRoles, id string) bool {
+	hasChanges := false
+
+	item := ResourceSMAdminRolesItems{}
+	for i := range data.Items {
+		if data.Items[i].Id.ValueString() == id {
+			item = data.Items[i]
+			break
+		}
+	}
+	stateItem := ResourceSMAdminRolesItems{}
+	for i := range state.Items {
+		if state.Items[i].Id.ValueString() == id {
+			stateItem = state.Items[i]
+			break
+		}
+	}
+	if !item.Name.Equal(stateItem.Name) {
+		hasChanges = true
+	}
+	if !item.Scope.Equal(stateItem.Scope) {
+		hasChanges = true
+	}
+	if !item.Tags.Equal(stateItem.Tags) {
+		hasChanges = true
+	}
+	return hasChanges
+}
+
+// End of section. //template:end hasChanges