+++ docs/guides/changelog.md
@@ -19,6 +19,7 @@ description: |-
 - Add `meraki_organization_policy_object_groups` resource
 - Add `meraki_switch_routing_interfaces` resource
 - Add `meraki_switch_stack_routing_interfaces` resource
+- Add `meraki_sm_admin_roles` resource
 
 ## 1.3.0